% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugResistantEpilepsy.R
\docType{package}
\name{DrugResistantEpilepsy-package}
\alias{DrugResistantEpilepsy}
\alias{DrugResistantEpilepsy-package}
\title{DrugResistantEpilepsy: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
